refactor challenge response
